requirements.txt
absl-py==2.0.0
active-semi-supervised-clustering==0.0.1
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.15.8
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
avalanche-lib==0.4.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
continuum==1.2.7
contourpy==1.1.1
cycler==0.12.0
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
docker-pycreds==0.4.0
dodgy==0.2.1
einops==0.7.0
ete3==3.1.3
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
flake8==5.0.4
flake8-polyfill==1.0.2
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.2
google-auth-oauthlib==1.0.0
GPUtil==1.4.0
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.18.0
idna==3.4
ImageHash==4.3.1
imageio==2.31.6
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
kneebow==1.0.1
kneed==0.8.5
lazy-object-proxy==1.9.0
lazy_loader==0.3
lightning-utilities==0.9.0
llvmlite==0.41.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
metric-learn==0.6.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy==1.6.1
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.9.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
numba==0.58.0
numpy==1.25.2
oauthlib==3.2.2
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pep8-naming==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.17.1
promise==2.3
prompt-toolkit==3.0.39
prospector==1.10.3
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.3.0
pydpp==0.2.1
pyflakes==2.5.0
Pygments==2.16.1
pylint==2.17.7
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pynndescent==0.5.10
pyparsing==3.1.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.4.2
pytest-mock==3.12.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorchcv==0.0.67
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
quadprog==0.1.11
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requirements-detector==1.2.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
self-attention-cv==1.2.3
semver==3.0.2
Send2Trash==1.8.2
sentry-sdk==1.31.0
setoptconf-tmp==0.3.1
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
stack-data==0.6.3
tbb==2021.10.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.4.9
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics==1.2.0
torchvision==0.12.0+cu113
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
types-python-dateutil==2.8.19.14
typing_extensions==4.4.0
tzdata==2023.3
umap-learn==0.5.4
uri-template==1.3.0
urllib3==2.0.6
visdom==0.2.4
wandb==0.15.12
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
yellowbrick==1.5
zipp==3.17.0
